pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slack-sql-qa-bot"
version = "0.1.0"
description = "Slack Q&A chatbot grounded in SQLite with LangGraph"
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "slack-bolt>=1.21.0",
  "aiohttp>=3.10.10",
  "langgraph>=0.2.35",
  "langgraph-checkpoint-sqlite>=3.0.0",
  "langchain-openai>=0.2.0",
  "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools.packages.find]
include = ["app*"]
